pyproject.toml
[tool.poetry]
name = "versioning"
version = "0.1.0"
description = ""
authors = ["Mahmud Youssef <mahmud.youssef@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.4"
ipython = "^8.17.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
